requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
euriai

sentence-transformers==2.2.2
transformers==4.39.3
huggingface_hub==0.19.4

langchain==0.1.13
langchain-community==0.0.29
langchain-text-splitters==0.0.1

pinecone==5.4.0
openai==1.23.6

pymupdf==1.23.25
python-docx==1.1.0
docx2txt==0.8

unstructured==0.12.5
unstructured[pdf]==0.12.5
unstructured[image]==0.12.5

pytesseract==0.3.10
opencv-python-headless==4.5.5.64  # safe for Docker


python-dotenv==1.0.1
scikit-learn==1.4.1.post1

faiss-cpu==1.8.0
rank-bm25==0.2.2

mlflow==2.10.2

pytest==8.2.0
requests==2.31.0

streamlit==1.33.0
sqlalchemy 
asyncpg 
psycopg2-binary

